refactor(View): split for readability
