+++ src/core-ts/View/index.tsx
@@ -1,26 +1,22 @@
-import vtkRenderer from '@kitware/vtk.js/Rendering/Core/Renderer';
-import vtkRenderWindow from '@kitware/vtk.js/Rendering/Core/RenderWindow';
-import vtkRenderWindowInteractor from '@kitware/vtk.js/Rendering/Core/RenderWindowInteractor';
-import vtkOpenGLRenderWindow from '@kitware/vtk.js/Rendering/OpenGL/RenderWindow';
 import { Nullable, Vector3 } from '@kitware/vtk.js/types';
 import {
   CSSProperties,
   forwardRef,
   PropsWithChildren,
-  useCallback,
   useEffect,
   useImperativeHandle,
   useMemo,
   useRef,
   useState,
 } from 'react';
-import { IView } from '../types';
-import useGetterRef from '../utils-ts/useGetterRef';
-import useMount from '../utils-ts/useMount';
-import { useOrderedUnmountContext } from '../utils-ts/useOrderedUnmountEffect';
-import useResizeObserver from '../utils-ts/useResizeObserver';
-import useUnmount from '../utils-ts/useUnmount';
-import { ViewContext } from './contexts';
+import { IView } from '../../types';
+import { useOrderedUnmountContext } from '../../utils-ts/useOrderedUnmountEffect';
+import { ViewContext } from '../contexts';
+import useInteractor from './useInteractor';
+import useRenderer from './useRenderer';
+import useRenderWindow from './useRenderWindow';
+import useRenderWindowView from './useRenderWindowView';
+import { useViewResize } from './useViewResize';
 
 const RENDERER_STYLE: CSSProperties = {
   position: 'absolute',
@@ -164,147 +160,37 @@ interface Props extends PropsWithChildren {
 const DefaultProps = {
   interactive: true,
   autoResetCamera: true,
-  background: [0.2, 0.3, 0.4],
+  background: [0.2, 0.3, 0.4] as Vector3,
   style: {
     width: '100%',
     height: '100%',
   } as CSSProperties,
 };
 
-function useRenderWindowView(container: Nullable<HTMLElement>) {
-  const [viewRef, getView] = useGetterRef(() => {
-    return vtkOpenGLRenderWindow.newInstance();
-  });
-
-  useEffect(() => {
-    const view = getView();
-    // FIXME setContainer API should allow null
-    view.setContainer(container as HTMLElement);
-  }, [container, getView]);
-
-  useUnmount(() => {
-    if (viewRef.current) {
-      viewRef.current.delete();
-      viewRef.current = null;
-    }
-  });
-
-  return getView;
-}
-
-function useInteractor(
-  getRWView: () => vtkOpenGLRenderWindow,
-  container: Nullable<HTMLElement>,
-  props: Props
-) {
-  const [interactorRef, getInteractor] = useGetterRef(() => {
-    return vtkRenderWindowInteractor.newInstance();
-  });
-  const { interactive = DefaultProps.interactive } = props;
-
-  useEffect(() => {
-    if (!container || !interactive) return;
-    const interactor = getInteractor();
-    const rwView = getRWView();
-
-    interactor.setView(rwView);
-    interactor.initialize();
-    interactor.bindEvents(container);
-  }, [interactive, container, getRWView, getInteractor]);
-
-  useUnmount(() => {
-    if (interactorRef.current) {
-      interactorRef.current.delete();
-      interactorRef.current = null;
-    }
-  });
-
-  return getInteractor;
-}
-
-function useRenderWindow(
-  getRWView: () => vtkOpenGLRenderWindow,
-  getRenderer: () => vtkRenderer
-) {
-  const [rwRef, getRenderWindow] = useGetterRef(() => {
-    return vtkRenderWindow.newInstance();
-  });
-
-  useEffect(() => {
-    const rwView = getRWView();
-    const renderWindow = getRenderWindow();
-    renderWindow.addView(rwView);
-    return () => {
-      renderWindow.removeView(rwView);
-    };
-  }, [getRWView, getRenderWindow]);
-
-  useEffect(() => {
-    const renderWindow = getRenderWindow();
-    const renderer = getRenderer();
-    renderWindow.addRenderer(renderer);
-    return () => {
-      renderWindow.removeRenderer(renderer);
-    };
-  }, [getRenderer, getRenderWindow]);
-
-  useUnmount(() => {
-    if (rwRef.current) {
-      rwRef.current.delete();
-      rwRef.current = null;
-    }
-  });
-
-  return getRenderWindow;
-}
-
-function useRenderer(props: Props) {
-  const [renRef, getRenderer] = useGetterRef(() => vtkRenderer.newInstance());
-  const { background = DefaultProps.background } = props;
-
-  useEffect(() => {
-    const renderer = getRenderer();
-    renderer.setBackground(background);
-  }, [background, getRenderer]);
-
-  useUnmount(() => {
-    if (renRef.current) {
-      renRef.current.delete();
-      renRef.current = null;
-    }
-  });
-
-  return getRenderer;
-}
-
 export default forwardRef(function View(props: Props, fwdRef) {
   const OrderedUnmountContext = useOrderedUnmountContext();
+
+  const {
+    background = DefaultProps.background,
+    interactive = DefaultProps.interactive,
+    autoResetCamera = DefaultProps.autoResetCamera,
+  } = props;
 
   const containerRef = useRef<Nullable<HTMLDivElement>>(null);
 
   const getRWView = useRenderWindowView(containerRef.current);
-  const getRenderer = useRenderer(props);
+  const getRenderer = useRenderer(background);
   // We need to attach the renderWindowView to the renderWindow
   // before setting the view onto the interactor.
   const getRenderWindow = useRenderWindow(getRWView, getRenderer);
-  const getInteractor = useInteractor(getRWView, containerRef.current, props);
-
-  const updateViewSize = useCallback(() => {
-    const container = containerRef.current;
-    if (!container) return;
-
-    const renderWindowView = getRWView();
-    const renderWindow = getRenderWindow();
-    const devicePixelRatio = window.devicePixelRatio || 1;
-    const { width, height } = container.getBoundingClientRect();
-    const w = Math.floor(width * devicePixelRatio);
-    const h = Math.floor(height * devicePixelRatio);
-    renderWindowView.setSize(Math.max(w, 10), Math.max(h, 10));
-    renderWindow.render();
-  }, [getRWView, getRenderWindow]);
+  const getInteractor = useInteractor(
+    getRWView,
+    containerRef.current,
+    interactive
+  );
 
   const [renderRequested, setRenderRequested] = useState(false);
-  const { autoResetCamera = DefaultProps.autoResetCamera } = props;
+  const requestRender = () => setRenderRequested(true);
 
   useEffect(() => {
     if (renderRequested) {
@@ -329,15 +215,13 @@ export default forwardRef(function View(props: Props, fwdRef) {
         This will batch render requests, triggering a single
        * vtk.js render once after a react render.
        */
-      requestRender: () => {
-        setRenderRequested(true);
-      },
+      requestRender,
       /**
        * Resets the camera.
        */
       resetCamera: () => {
         getRenderer().resetCamera();
-        getRenderWindow().render();
+        requestRender();
       },
     };
   }, [getRWView, getRenderer, getRenderWindow, getInteractor]);
@@ -345,8 +229,7 @@ export default forwardRef(function View(props: Props, fwdRef) {
   // expose the view as a ref for imperative control
   useImperativeHandle(fwdRef, () => view);
 
-  useResizeObserver(containerRef.current, updateViewSize);
-  useMount(() => updateViewSize());
+  useViewResize(containerRef, view);
 
   const { style = DefaultProps.style } = props;
   const containerStyle = useMemo<CSSProperties>(
